requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pydantic>=2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
